+++ apps/companion/src/app/run-filters.tsx
@@ -1,0 +1,3 @@
+import { RunFiltersScreen } from '../components/RunFiltersScreen';
+
+export default RunFiltersScreen;